Made the share button reusable component and added to editor page
