+++ floor-plan/src/app/components/Modal.module.css
@@ -8,6 +8,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
+    z-index: 1000;
 }
 
 .modal {
@@ -24,7 +25,7 @@
     margin-top: 20px;
 }
 
-.button {
+.modalButton {
     padding: 10px 20px;
     border: none;
     background-color: #007bff;
@@ -34,7 +35,7 @@
     font-size: 14px;
 }
 
-.button:hover {
+.modalButton:hover {
     background-color: #0056b3;
 }
