+++ floor-plan/src/app/components/Modal.tsx
@@ -1,6 +1,7 @@
 // Reusable and custom modal 
 import React from 'react';
 import styles from './Modal.module.css'; 
+import ReactDOM from "react-dom";
 
 interface ModalProps {
     isVisible: boolean;
@@ -13,20 +14,20 @@ interface ModalProps {
 const Modal: React.FC<ModalProps> = ({ isVisible, onClose, onConfirm, children, title }) => {
     if (!isVisible) return null; // Don't render modal if it's not visible
 
-    return (
+    return ReactDOM.createPortal(
         <div className={styles.modalOverlay}>
             <div className={styles.modal}>
-                {title && <h2>{title}</h2>}
-                <div className={styles.modalContent}>
-                    {children}
-                </div>
+                <h2>{title}</h2>
+                <div className={styles.modalContent}>{children}</div>
                 <div className={styles.modalButtons}>
-                    <button className={styles.button} onClick={onConfirm}>Confirm</button>
-                    <button className={styles.button} onClick={onClose}>Cancel</button>
+                    <button onClick={onConfirm} className={styles.modalButton}>Confirm</button>
+                    <button onClick={onClose} className={styles.modalButton}>Cancel</button>
                 </div>
             </div>
-        </div>
+        </div>,
+        document.body // This renders the modal directly in the body
     );
 };
+
 
 export default Modal;
