+++ floor-plan/src/app/home/page.module.css
@@ -218,6 +218,7 @@
 }
 
 .fileItem {
+  position: relative;
   display: flex;
   flex-direction: column;
   align-items: center;
